requirements.txt
asgiref==3.5.2
Django==4.0.4
django-cors-headers==3.14.0
django-resized==1.0.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
Pillow==9.5.0
psycopg2-binary
PyJWT==2.6.0
pytz==2023.3
sqlparse==0.4.2
whitenoise==6.2.0
